+++ CHANGELOG.md
@@ -35,6 +35,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - bugfix ``Regex#replace`` method when replce with empty string
+- not show error line marker when reach EOS and previous token size is 1
 
 
 ## [0.20.0] - 2020-12-31